Refactor Genes/Mutations to use new hooks create hooks in genomic add filter.json to Genes/Mutation app
