+++ packages/portal-proto/src/features/facets/hooks.ts
@@ -1,6 +1,4 @@
 import {
-  CoreState,
-  CoreDispatch,
   Operation,
   EnumValueExtractorHandler,
   EnumOperandValue,
@@ -9,16 +7,12 @@ import {
   handleOperation,
   FilterSet,
   removeCohortFilter,
-  removeGenomicFilter,
   selectCurrentCohortFilters,
   selectCurrentCohortFiltersByName,
   fetchFacetByNameGQL,
   updateCohortFilter,
-  updateGenomicFilter,
   useCoreDispatch,
   useCoreSelector,
-  selectGenomicFilters,
-  selectGenomicFiltersByName,
   GQLDocType,
   GQLIndexType,
   NumericFromTo,
@@ -26,8 +20,6 @@ import {
   fetchFacetContinuousAggregation,
   selectFacetByDocTypeAndField,
   usePrevious,
-  selectGenomicAndCohortFilters,
-  selectCurrentCohortFilterOrCaseSet,
   selectTotalCountsByName,
 } from "@gff/core";
 import { useEffect } from "react";
@@ -39,22 +31,12 @@ import {
   UpdateFacetFilterFunction,
 } from "@/features/facets/types";
 import { AnyAction, ThunkDispatch } from "@reduxjs/toolkit";
-import { TypedUseSelectorHook } from "react-redux";
-import { partial } from "lodash";
 
 /**
  * Filter selector for all the facet filters
  */
 const useCohortFacetFilter = (): FilterSet => {
   return useCoreSelector((state) => selectCurrentCohortFilters(state));
-};
-
-const useCohortOrCaseSetFacetFilter = (): FilterSet => {
-  return useCoreSelector((state) => selectCurrentCohortFilterOrCaseSet(state));
-};
-
-const useGenomicFacetFilter = (): FilterSet => {
-  return useCoreSelector((state) => selectGenomicFilters(state));
 };
 
 export const extractValue = (op: Operation): EnumOperandValue => {
@@ -74,21 +56,14 @@ const useCohortFacetFilterByName = (field: string): OperandValue => {
   return enumFilters ? extractValue(enumFilters) : undefined;
 };
 
-const useGenomicFilterByName = (field: string): OperandValue => {
-  const enumFilters: Operation = useCoreSelector((state) =>
-    selectGenomicFiltersByName(state, field),
-  );
-  return enumFilters ? extractValue(enumFilters) : undefined;
-};
-
 /**
  *  Facet Selector using GQL which will refresh when filters/enum values changes.
  */
 
 export const useEnumFacet = (
+  field: string,
   docType: GQLDocType,
   indexType: GQLIndexType,
-  field: string,
 ): EnumFacetResponse => {
   const coreDispatch = useCoreDispatch();
 
@@ -138,129 +113,6 @@ export const useEnumFacet = (
   };
 };
 
-/**
- * Genes Facet Selector using GQL. it combines the Cohort with Gene Filters
- * to get data for the current cohort and genes filters
- */
-export const useGenesFacet = (
-  field: string,
-  docType = "genes" as GQLDocType,
-  indexType = "explore" as GQLIndexType,
-): EnumFacetResponse => {
-  const coreDispatch = useCoreDispatch();
-  const facet: FacetBuckets = useCoreSelector((state) =>
-    selectFacetByDocTypeAndField(state, docType, field),
-  );
-
-  const enumValues = useGenomicFilterByName(field);
-  const cohortFilters = useCohortOrCaseSetFacetFilter();
-  const genomicFilters = useGenomicFacetFilter();
-  const prevCohortFilters = usePrevious(cohortFilters);
-  const prevGenomicFilters = usePrevious(genomicFilters);
-  const prevEnumValues = usePrevious(enumValues);
-
-  useEffect(() => {
-    if (
-      !facet ||
-      !isEqual(prevCohortFilters, cohortFilters) ||
-      !isEqual(prevGenomicFilters, genomicFilters) ||
-      !isEqual(prevEnumValues, enumValues)
-    ) {
-      coreDispatch(
-        fetchFacetByNameGQL({
-          field: field,
-          docType: docType,
-          index: indexType,
-          filterSelector: selectGenomicAndCohortFilters,
-        }),
-      );
-    }
-  }, [
-    coreDispatch,
-    facet,
-    field,
-    cohortFilters,
-    docType,
-    indexType,
-    prevCohortFilters,
-    prevEnumValues,
-    enumValues,
-    prevGenomicFilters,
-    genomicFilters,
-  ]);
-
-  return {
-    data: facet?.buckets,
-    enumFilters: (enumValues as EnumOperandValue)?.map((x) => x.toString()),
-    error: facet?.error,
-    isUninitialized: facet === undefined,
-    isFetching: facet?.status === "pending",
-    isSuccess: facet?.status === "fulfilled",
-    isError: facet?.status === "rejected",
-  };
-};
-
-/**
- * Mutations Facet Selector using GQL
- */
-const useMutationsFacet = (
-  field: string,
-  docType = "ssms" as GQLDocType,
-  indexType = "explore" as GQLIndexType,
-): EnumFacetResponse => {
-  const coreDispatch = useCoreDispatch();
-  const facet: FacetBuckets = useCoreSelector((state) =>
-    selectFacetByDocTypeAndField(state, docType, field),
-  );
-
-  const enumValues = useGenomicFilterByName(field);
-  const cohortFilters = useCohortOrCaseSetFacetFilter();
-  const genomicFilters = useGenomicFacetFilter();
-  const prevCohortFilters = usePrevious(cohortFilters);
-  const prevGenomicFilters = usePrevious(genomicFilters);
-  const prevEnumValues = usePrevious(enumValues);
-
-  useEffect(() => {
-    if (
-      !facet ||
-      !isEqual(prevCohortFilters, cohortFilters) ||
-      !isEqual(prevGenomicFilters, genomicFilters) ||
-      !isEqual(prevEnumValues, enumValues)
-    ) {
-      coreDispatch(
-        fetchFacetByNameGQL({
-          field: field,
-          docType: docType,
-          index: indexType,
-          filterSelector: selectGenomicAndCohortFilters,
-        }),
-      );
-    }
-  }, [
-    coreDispatch,
-    facet,
-    field,
-    cohortFilters,
-    docType,
-    indexType,
-    prevCohortFilters,
-    prevEnumValues,
-    enumValues,
-    prevGenomicFilters,
-    genomicFilters,
-  ]);
-
-  return {
-    data: facet?.buckets,
-    enumFilters: (enumValues as EnumOperandValue)?.map((x) => x.toString()),
-    error: facet?.error,
-    isUninitialized: facet === undefined,
-    isFetching: facet?.status === "pending",
-    isSuccess: facet?.status === "fulfilled",
-    isError: facet?.status === "rejected",
-  };
-};
-
 type UpdateEnumFiltersFunc = (
   dispatch: ThunkDispatch<any, undefined, AnyAction>,
   field: string,
@@ -297,42 +149,11 @@ export const updateEnumFilters: UpdateEnumFiltersFunc = (
   }
 };
 
-// Hook version of updateEnumFilters
-
-export const useUpdateEnumFilters = () => {
-  const dispatch = useCoreDispatch();
-  return (field: string, enumerationFilters: EnumOperandValue) =>
-    partial(updateEnumFilters, dispatch, field, enumerationFilters);
-};
-
-export const useUpdateGenomicEnumFilters: UpdateEnumFiltersFunc = (
-  dispatch: ThunkDispatch<any, undefined, AnyAction>,
-  field: string,
-  enumerationFilters: EnumOperandValue,
-) => {
-  if (enumerationFilters === undefined) dispatch(removeGenomicFilter(field));
-  if (enumerationFilters.length > 0) {
-    dispatch(
-      updateGenomicFilter({
-        field: field,
-        operation: {
-          operator: "includes",
-          field: field,
-          operands: enumerationFilters,
-        },
-      }),
-    );
-  } else {
-    // completely remove the field
-    dispatch(removeGenomicFilter(field));
-  }
-};
-
 export const useRangeFacet = (
-  docType: GQLDocType,
-  indexType: GQLIndexType,
   field: string,
   ranges: ReadonlyArray<NumericFromTo>,
+  docType: GQLDocType,
+  indexType: GQLIndexType,
 ): FacetResponse => {
   const coreDispatch = useCoreDispatch();
   const facet: FacetBuckets = useCoreSelector((state) =>
@@ -388,25 +209,6 @@ export const useSelectFieldFilter = (field: string): Operation => {
   );
 };
 
-// Global Selector for Facet Values
-export const selectFieldFilter = (
-  selector: TypedUseSelectorHook<CoreState>,
-  field: string,
-): Operation => {
-  // get the current filter for this facet
-  return selector((state) => selectCurrentCohortFiltersByName(state, field));
-};
-
-// Update Core Filter
-export const updateFacetFilter = (
-  dispatch: CoreDispatch,
-  field: string,
-  operation: Operation,
-): void => {
-  // update the filter for this facet
-  dispatch(updateCohortFilter({ field: field, operation: operation }));
-};
-
 // Update filter hook
 export const useUpdateFacetFilter = (): UpdateFacetFilterFunction => {
   const dispatch = useCoreDispatch();
@@ -414,11 +216,6 @@ export const useUpdateFacetFilter = (): UpdateFacetFilterFunction => {
   return (field: string, operation: Operation) => {
     dispatch(updateCohortFilter({ field: field, operation: operation }));
   };
-};
-
-// Global Clear
-export const clearFilters = (dispatch: CoreDispatch, field: string): void => {
-  dispatch(removeCohortFilter(field));
 };
 
 // Core ClearFilters hook
@@ -431,20 +228,6 @@ export const useClearFilters = (): ClearFacetFunction => {
 
 export const useTotalCounts = (name: string): number =>
   useCoreSelector((state) => selectTotalCountsByName(state, name));
-
-export const UpdateEnums = {
-  cases: updateEnumFilters,
-  files: updateEnumFilters,
-  genes: useUpdateGenomicEnumFilters,
-  ssms: useUpdateGenomicEnumFilters,
-};
-
-export const FacetEnumHooks = {
-  cases: useEnumFacet,
-  files: useEnumFacet,
-  genes: useGenesFacet,
-  ssms: useMutationsFacet,
-};
 
 export const FacetDocTypeToCountsIndexMap = {
   cases: "caseCounts",
