+++ packages/portal-proto/src/features/oncoGrid/filters.json
@@ -3,34 +3,40 @@
     {
       "facet_filter": "genes.biotype",
       "name": "Biotype",
-      "description": "No description"
+      "description": "No description",
+      "type": "enum"
     },
     {
       "facet_filter": "genes.is_cancer_gene_census",
       "name": "Is Cancer Gene Census",
-      "description": "No description"
+      "description": "No description",
+      "type": "toggle"
     }
   ],
   "ssms": [
     {
       "facet_filter": "ssms.consequence.transcript.annotation.vep_impact",
       "name": "VEP Impact",
-      "description": "No description"
+      "description": "No description",
+      "type": "enum"
     },
     {
       "facet_filter": "ssms.consequence.transcript.annotation.sift_impact",
       "name": "SIFT Impact",
-      "description": "No description"
+      "description": "No description",
+      "type": "enum"
     },
     {
       "facet_filter": "ssms.consequence.transcript.annotation.polyphen_impact",
       "name": "Polyphen Impact",
-      "description": "No description"
+      "description": "No description",
+      "type": "enum"
     },
     {
       "facet_filter": "ssms.mutation_subtype",
       "name": "Type",
-      "description": "No description"
+      "description": "No description",
+      "type": "enum"
     }
   ]
 }
